+++ src/com/upokecenter/text/ByteData.java
@@ -23,10 +23,10 @@ import com.upokecenter.util.*;
      */
     public boolean GetBoolean(int cp) {
       if (cp < 0) {
-        throw new IllegalArgumentException("cp (" + Long.toString((long)cp) + ") is less than " + "0");
+        throw new IllegalArgumentException("cp (" + Integer.toString((int)cp) + ") is less than " + "0");
       }
       if (cp > 0x10ffff) {
-        throw new IllegalArgumentException("cp (" + Long.toString((long)cp) + ") is more than " + Long.toString((long)0x10ffff));
+        throw new IllegalArgumentException("cp (" + Integer.toString((int)cp) + ") is more than " + Integer.toString((int)0x10ffff));
       }
       int b = this.array[cp >> 13] & 0xff;
       switch (b) {
@@ -50,10 +50,10 @@ import com.upokecenter.util.*;
      */
     public byte GetByte(int cp) {
       if (cp < 0) {
-        throw new IllegalArgumentException("cp (" + Long.toString((long)cp) + ") is less than " + "0");
+        throw new IllegalArgumentException("cp (" + Integer.toString((int)cp) + ") is less than " + "0");
       }
       if (cp > 0x10ffff) {
-        throw new IllegalArgumentException("cp (" + Long.toString((long)cp) + ") is more than " + Long.toString((long)0x10ffff));
+        throw new IllegalArgumentException("cp (" + Integer.toString((int)cp) + ") is more than " + Integer.toString((int)0x10ffff));
       }
       int index = (cp >> 9) << 1;
       int x = this.array[index + 1];
