+++ src/com/upokecenter/text/IntArrayCharacterInput.java
@@ -35,19 +35,19 @@ package com.upokecenter.text;
         throw new NullPointerException("buf");
       }
       if (offset < 0) {
-        throw new IllegalArgumentException("offset (" + Long.toString((long)offset) + ") is less than " + "0");
+        throw new IllegalArgumentException("offset (" + Integer.toString((int)offset) + ") is less than " + "0");
       }
       if (offset > buf.length) {
-        throw new IllegalArgumentException("offset (" + Long.toString((long)offset) + ") is more than " + Long.toString((long)buf.length));
+        throw new IllegalArgumentException("offset (" + Integer.toString((int)offset) + ") is more than " + Integer.toString((int)buf.length));
       }
       if (unitCount < 0) {
-        throw new IllegalArgumentException("unitCount (" + Long.toString((long)unitCount) + ") is less than " + "0");
+        throw new IllegalArgumentException("unitCount (" + Integer.toString((int)unitCount) + ") is less than " + "0");
       }
       if (unitCount > buf.length) {
-        throw new IllegalArgumentException("unitCount (" + Long.toString((long)unitCount) + ") is more than " + Long.toString((long)buf.length));
+        throw new IllegalArgumentException("unitCount (" + Integer.toString((int)unitCount) + ") is more than " + Integer.toString((int)buf.length));
       }
       if (buf.length - offset < unitCount) {
-        throw new IllegalArgumentException("buf's length minus " + offset + " (" + Long.toString((long)(buf.length - offset)) + ") is less than " + Long.toString((long)unitCount));
+        throw new IllegalArgumentException("buf's length minus " + offset + " (" + Integer.toString((int)(buf.length - offset)) + ") is less than " + Integer.toString((int)unitCount));
       }
       if (unitCount == 0) {
         return 0;
