+++ src/com/upokecenter/text/PartialListCharacterInput.java
@@ -15,19 +15,19 @@ import java.util.*;
         throw new NullPointerException("ilist");
       }
       if (start < 0) {
-        throw new IllegalArgumentException("start (" + Long.toString((long)start) + ") is less than " + "0");
+        throw new IllegalArgumentException("start (" + Integer.toString((int)start) + ") is less than " + "0");
       }
       if (start > ilist.size()) {
-        throw new IllegalArgumentException("start (" + Long.toString((long)start) + ") is more than " + Long.toString((long)ilist.size()));
+        throw new IllegalArgumentException("start (" + Integer.toString((int)start) + ") is more than " + Integer.toString((int)ilist.size()));
       }
       if (length < 0) {
-        throw new IllegalArgumentException("length (" + Long.toString((long)length) + ") is less than " + "0");
+        throw new IllegalArgumentException("length (" + Integer.toString((int)length) + ") is less than " + "0");
       }
       if (length > ilist.size()) {
-        throw new IllegalArgumentException("length (" + Long.toString((long)length) + ") is more than " + Long.toString((long)ilist.size()));
+        throw new IllegalArgumentException("length (" + Integer.toString((int)length) + ") is more than " + Integer.toString((int)ilist.size()));
       }
       if (ilist.size() - start < length) {
-        throw new IllegalArgumentException("ilist's length minus " + start + " (" + Long.toString((long)(ilist.size() - start)) + ") is less than " + Long.toString((long)length));
+        throw new IllegalArgumentException("ilist's length minus " + start + " (" + Integer.toString((int)(ilist.size() - start)) + ") is less than " + Integer.toString((int)length));
       }
       this.ilist = ilist;
       this.pos = start;
@@ -57,19 +57,19 @@ import java.util.*;
         throw new NullPointerException("buf");
       }
       if (offset < 0) {
-        throw new IllegalArgumentException("offset (" + Long.toString((long)offset) + ") is less than " + "0");
+        throw new IllegalArgumentException("offset (" + Integer.toString((int)offset) + ") is less than " + "0");
       }
       if (offset > buf.length) {
-        throw new IllegalArgumentException("offset (" + Long.toString((long)offset) + ") is more than " + Long.toString((long)buf.length));
+        throw new IllegalArgumentException("offset (" + Integer.toString((int)offset) + ") is more than " + Integer.toString((int)buf.length));
       }
       if (unitCount < 0) {
-        throw new IllegalArgumentException("unitCount (" + Long.toString((long)unitCount) + ") is less than " + "0");
+        throw new IllegalArgumentException("unitCount (" + Integer.toString((int)unitCount) + ") is less than " + "0");
       }
       if (unitCount > buf.length) {
-        throw new IllegalArgumentException("unitCount (" + Long.toString((long)unitCount) + ") is more than " + Long.toString((long)buf.length));
+        throw new IllegalArgumentException("unitCount (" + Integer.toString((int)unitCount) + ") is more than " + Integer.toString((int)buf.length));
       }
       if (buf.length - offset < unitCount) {
-        throw new IllegalArgumentException("buf's length minus " + offset + " (" + Long.toString((long)(buf.length - offset)) + ") is less than " + Long.toString((long)unitCount));
+        throw new IllegalArgumentException("buf's length minus " + offset + " (" + Integer.toString((int)(buf.length - offset)) + ") is less than " + Integer.toString((int)unitCount));
       }
       if (unitCount == 0) {
         return 0;
